WhatsApp Marketing Messages ROI Assessment
The goal of this script is to provide technical detail on a proposal to define how Power Calculation and Data Analysis should be defined when piloting WhatsApp Marketing Messages measurement with clients. This is a new measurement proposal must be executed at the advertiser / partner side, in their environment.
It's used to compare WhatsApp Marketing Messages performance against other channels such as Email or SMS Marketing.

The method consists of using conversion rates estimates, applying the fixed cost to the amount of impressions and calculate CPA.